# api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAHWI Career Expo API Documentation",
        "description": "Comprehensive API documentation for the SAHWI Career Expo platform. This API provides endpoints for event management, user authentication, registrations, payments, messaging, forums, articles, and more.",
        "version": "1.0.0",
        "contact": {
            "email": "support@sahwi-career-expo.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://sahwi-career-expo.com/license"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Enter your bearer token in the format: Bearer {token}"
            }
        }
    },
...
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Registration",
                "description": "Register a new user account. Different roles require different fields. Students are auto-verified, while other roles require admin approval.",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password",
                                    "name",
                                    "role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "newuser@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePass123!"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "student",
                                            "professional",
                                            "company_rep",
                                            "university",
                                            "single"
                                        ],
                                        "example": "student"
                                    },
                                    "current_school_name": {
                                        "type": "string",
                                        "example": "ABC High School",
                                        "description": "Required for students"
                                    },
                                    "current_grade": {
                                        "type": "string",
                                        "example": "Grade 12",
                                        "description": "Required for students"
                                    },
                                    "interested_area": {
                                        "type": "string",
                                        "example": "Computer Science",
                                        "description": "Required for students"
                                    },
                                    "interested_course": {
                                        "type": "string",
                                        "example": "BSc Computer Science",
                                        "description": "Required for students"
                                    },
                                    "interested_university_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "description": "Required for students"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration successful. Please check your email to verify your account."
                                        },
                                        "user": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "newuser@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "student"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}